Initialized duzenlendi ve build oncesine tasindi
